backend/requirements.txt
fastapi==0.115.0
uvicorn==0.31.0
chromadb==0.5.0
openai==1.3.5
sentence-transformers==3.0.1
pydantic==2.8.0
python-multipart==0.0.12
typing-extensions>=4.12.2
numpy<2.0
python-dotenv==1.0.0
